finsight-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
pandas==2.1.3
python-multipart==0.0.6
httpx==0.25.1